Client now communicating with Broker
Client is now sending commands to the Mqtt broker on the raspberry pi.
When the client is running the user has the option of turning the light
on, off or shutting down the connection.
Currently these commands are being sent as strings (i.e On Command or
Off Command). This will be changed to deliver protobuf payloads.

This was implemented using the following patterns

- Command pattern
- Singleton pattern